refactor: chat user auth ui
