+++ ui/src/views/system-chat-user/authentication/component/OAuth2.vue
@@ -100,7 +100,7 @@
 </template>
 <script setup lang="ts">
 import { reactive, ref, onMounted } from 'vue'
-import authApi from '@/api/system/auth'
+import authApi from '@/api/chat-user/auth-setting'
 import type { FormInstance, FormRules } from 'element-plus'
 import { t } from '@/locales'
 import { MsgSuccess } from '@/utils/message'
